chore: v0.3.12 — HP/MP int32 擴展 + 文字編碼可配置
- feat: HP/MP 上限擴展至 int32（最大 9,999,999），DB migration 027，封包改 WriteD
- feat: 客戶端文字編碼可配置化（server.toml client_language_code: MS950/GBK），新增 packet/encoding.go
- feat: 裝備欄位系統擴展（equipment.go 支援更多部位）+ .speed GM 指令
- feat: 新增測試用防具資料（armor_list.yaml）

+++ internal/component/character.go
@@ -11,10 +11,10 @@ type Character struct {
 
 	Level int16
 	Exp   int64
-	HP    int16
-	MP    int16
-	MaxHP int16
-	MaxMP int16
+	HP    int32
+	MP    int32
+	MaxHP int32
+	MaxMP int32
 	AC    int16
 
 	Str   int16
